requirements.txt
django==4.0.1
djangorestframework==3.13.1
djoser==2.1.0
black==21.12b0
pytest==6.2.5
gunicorn==20.1.0